requirements.txt
fastapi
uvicorn
langchain-openai
numpy
pydantic
openai
faiss-cpu
pillow
transformers
sqlalchemy
torch
pgvector
faiss
python-dotenv
psycopg